backend/pyproject.toml
[project]
name = "hackathon-agent-core-backend"
version = "0.1.0"
description = "Minimal Claude/Gemini chatbot backend with streaming + tool calls"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "greenlet>=3.0.0",
  "aiosqlite>=0.20.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "anthropic>=0.42.0",
  "litellm>=1.58.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
